add gearbox to helm (now includes PEDS-1378 and PEDS-1356)

+++ tools/roll.sh
@@ -5,11 +5,31 @@ cd "$(dirname "$0")/../helm/gen3" || exit 1 || exit 1
 
 rm ../../values.yaml
 
+# Initialize variables
+DISABLE_UPDATE=false
+
+
+
+project="$1"
+shift
+
+# Parse arguments
+for arg in "$@"; do
+  case $arg in
+    --disable-update)
+      DISABLE_UPDATE=true
+      shift # Remove --disable-update from arguments
+      ;;
+    *)
+      # Pass through other arguments
+      ;;
+  esac
+done
 # Check if ../../secret-values.yaml exists
 if [ -f ../../secret-values.yaml ]; then
-  yq '. *= load("../../secret-values.yaml")' ../../default-values.yaml > ../../values.yaml
+  yq '. *= load("../../secret-values.yaml")' ../../$project-default-values.yaml > ../../values.yaml
 else 
-  cp ../../default-values.yaml ../../values.yaml
+  cp ../../$project-default-values.yaml ../../values.yaml
 fi
 # Directory to store CA certificate
 ca_dir=../../CA
@@ -38,13 +58,26 @@ if [ $# -gt 0 ]; then
   # Iterate over each argument (service name)
   for service_name in "$@"
   do
+    # Skip if service_name is disable-update
+    if [ "$service_name" = "--disable-update" ]; then
+      continue
+    fi
+
     # Delete the deployment corresponding to the service name
     kubectl delete deployment ${service_name}-deployment
+    if [ "$service_name" = "gearbox" ]; then
+      kubectl delete job gearbox-g3auto-patch
+    fi
   done
 fi
 
-# Run helm dependency update
-helm dependency update
+# Conditionally run helm dependency update
+if [ "$DISABLE_UPDATE" = false ]; then
+  echo "Running helm dependency update..."
+  helm dependency update
+else
+  echo "Skipping helm dependency update..."
+fi
 
 # Run helm upgrade --install command
-helm upgrade --install pcdc . -f ../../values.yaml
+helm upgrade --install $project . -f ../../values.yaml
